+++ src/calculate.h
@@ -6,7 +6,7 @@
 int add_mat(Matrix *mat, int n);
 int sub_mat(Matrix *mat, int n);
 int mul_mat(Matrix *mat, int n);
-int det_mat(const Matrix *mat);
+double det_mat(const Matrix *mat);
 int _add_mat(Matrix *a, const Matrix *b);
 int _sub_mat(Matrix *a, const Matrix *b);
 int _mul_mat(Matrix *a, const Matrix *b);
